# src/main/resources/api-swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "Spring Boot REST API - MYAPP-DEMO",
    "version": "1.0",
    "title": "SPRING BOOT - MYAPP-DEMO",
    "termsOfService": "Terminos del servicio",
    "contact": {
      "name": "José Alfonso Arauco Villar - JAAV",
      "url": "http://localhost:8081",
      "email": "araucovillar@gmail.com"
    },
    "license": {
      "name": "Apache License Version 2.0",
      "url": "https://www.apache.org/licenses/LICENSE-2.0"
    }
  },
  "host": "localhost:8081",
  "basePath": "/",
  "tags": [
    {
      "name": "status-controller",
      "description": "Status Controller"
    },
    {
      "name": "procces-controller",
      "description": "Procces Controller"
    }
  ],
  "paths": {
    "/api/channel/sys/v1/procces/execute": {
      "post": {
        "tags": [
          "procces-controller"
        ],
        "summary": "Ejecutar proceso X",
        "operationId": "executeUsingPOST",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "type",
            "in": "query",
            "description": "type",
            "required": false,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "ok",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/ProccesEntity"
              }
            }
          },
          "201": {
            "description": "Created"
          },
          "400": {
            "description": "BAD RQ",
            "schema": {
              "$ref": "#/definitions/Exception"
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/api/status/time": {
      "get": {
        "tags": [
          "status-controller"
        ],
        "summary": "Verificar time",
        "operationId": "getTimeUsingGET",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "ok",
            "schema": {
              "type": "string"
            }
          },
          "400": {
            "description": "BAD RQ",
            "schema": {
              "$ref": "#/definitions/Exception"
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/api/status/verify": {
      "get": {
        "tags": [
          "status-controller"
        ],
        "summary": "Verificar estado API",
        "operationId": "verificarStatusUsingGET",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "ok",
            "schema": {
              "type": "string"
            }
          },
          "400": {
            "description": "BAD RQ",
            "schema": {
              "$ref": "#/definitions/Exception"
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    }
  },
  "definitions": {
    "Exception": {
      "type": "object",
      "properties": {
        "cause": {
          "$ref": "#/definitions/Throwable"
        },
        "localizedMessage": {
          "type": "string"
        },
        "message": {
          "type": "string"
        },
        "stackTrace": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/StackTraceElement"
          }
        },
        "suppressed": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Throwable"
          }
        }
      }
    },
    "ProccesEntity": {
      "type": "object",
      "properties": {
        "countError": {
          "type": "integer",
          "format": "int32"
        },
        "countOk": {
          "type": "integer",
          "format": "int32"
        },
        "proccesName": {
          "type": "string"
        },
        "status": {
          "type": "string"
        }
      }
    },
    "StackTraceElement": {
      "type": "object",
      "properties": {
        "className": {
          "type": "string"
        },
        "fileName": {
          "type": "string"
        },
        "lineNumber": {
          "type": "integer",
          "format": "int32"
        },
        "methodName": {
          "type": "string"
        },
        "nativeMethod": {
          "type": "boolean"
        }
      }
    },
    "Throwable": {
      "type": "object",
      "properties": {
        "cause": {
          "$ref": "#/definitions/Throwable"
        },
        "localizedMessage": {
          "type": "string"
        },
        "message": {
          "type": "string"
        },
        "stackTrace": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/StackTraceElement"
          }
        },
        "suppressed": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Throwable"
          }
        }
      }
    }
  }
}